# specs/openapi.yaml
openapi: 3.0.3
info:
  title: AI 卫星总体设计助手 API
  version: 0.4.0
  description: |
    由 `docs/07-api.md`（v0.2 细化版）生成，用途：**Apifox / Postman 直接导入进行手动测试**。
    - 本地 Base URL：`http://127.0.0.1:8000/api/v1`
    - 认证：`Authorization: Bearer <APP_TOKEN>`（仅 `/healthz` 除外）
    - 任务异步：创建返回 `202`；进度经 SSE（`GET /missions/{id}/events`）或轮询（`GET /missions/{id}`）
    - `STUB_LLM=1` 时为桩模式（离线演示 / CI，全部 LLM 调用走录制回放）
    - 枚举、错误码、SSE 事件全表见 `docs/07-api.md` §1
servers:
  - url: http://127.0.0.1:8000/api/v1
    description: 本地开发
...
security:
  - bearerAuth: []
paths:
...
  /missions/{id}/parameters:
    parameters:
      - $ref: '#/components/parameters/MissionId'
    get:
      operationId: listParameters
      tags: [Parameters]
      summary: 参数列表（A6）
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ParameterStatus' }
        - name: module
          in: query
          schema: { $ref: '#/components/schemas/Module' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterList'
              example:
                items:
                  - id: power.sa.area
                    name: 太阳翼面积
                    unit: m2
                    value: 4.8
                    status: verified
                    source_type: calc
                    margin: 0.1
                    validation_status: pass
                    updated_at: '2026-10-05T09:32:00Z'
                total: 1
                page: 1
                page_size: 50
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /missions/{id}/parameters/{pid}:
    parameters:
      - $ref: '#/components/parameters/MissionId'
      - name: pid
        in: path
        required: true
        schema: { type: string }
        example: power.sa.area
    get:
      operationId: getParameter
      tags: [Parameters]
      summary: 参数详情（A7）
      description: 完整参数对象 + provenance 溯源链 + 关联校验结论。
      responses:
        '200':
          description: 详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterDetail'
              example:
                id: power.sa.area
                name: 太阳翼面积
                value: 4.8
                unit: m2
                source_type: calc
                source_ref: calc-20261005-0007
                margin: 0.1
                confidence: high
                status: verified
                version: 3
                provenance:
                  type: calculation
                  record:
                    id: calc-20261005-0007
                    tool: calc.eps
                    formula_version: v1
                    inputs: { power.p_eol_w: 210.5, cell_efficiency: 0.3 }
                    outputs: { power.sa.area: 4.8 }
                    code_commit: a1b2c3d
                validations:
                  - { rule_id: V4, status: pass, created_at: '2026-10-05T09:03:00Z' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /missions/{id}/parameters/{pid}/review:
    parameters:
      - $ref: '#/components/parameters/MissionId'
      - name: pid
        in: path
        required: true
        schema: { type: string }
        example: mass.margin_pct
    post:
      operationId: reviewParameter
      tags: [Parameters]
      summary: 参数评审（A8，人工门禁）
      description: |
        业务逻辑：① 参数须为 needs_review / proposed（否则 409）；
        ② accept → verified（可带 edited_value 先更新数值并版本 +1）；reject → rejected（comment 必填）；
        ③ 写入 review_gates 与 audit_log；④ 自动触发 A9 重校验并推送 SSE。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
            example:
              action: accept
              comment: 按工程经验接受 10% 余量
              reviewer: engineer@example.com
      responses:
        '200':
          description: 评审完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
              example:
                param_id: mass.margin_pct
                status: verified
                gate: { gate_id: g-003, action: accept, reviewer: engineer@example.com, ts: '2026-10-05T09:05:00Z' }
                revalidation: { queued: true, scope: 'param:mass.margin_pct' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: APP_TOKEN
  parameters:
    MissionId:
      name: id
      in: path
      required: true
      schema: { type: string, pattern: '^t-\d{8}-\d{3}$' }
      example: t-20261005-001
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    BadRequest:
      description: 入参不符合 JSON Schema
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error:
              code: SCHEMA_INVALID
              message: constraints.lifetime_years 必须 ≥ 0.5
              details: { field: constraints.lifetime_years }
              trace_id: req-20261005-abc123
    Unauthorized:
      description: Token 缺失或错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: UNAUTHORIZED, message: 无效的访问令牌, details: {}, trace_id: req-20261005-abc124 }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: MISSION_NOT_FOUND, message: 任务不存在, details: { task_id: t-20261005-999 }, trace_id: req-20261005-abc125 }
    Conflict:
      description: 状态冲突（同 409 各错误码：DOC_NOT_READY / STATE_CONFLICT）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: STATE_CONFLICT, message: 当前任务状态不允许该操作, details: { status: succeeded }, trace_id: req-20261005-abc126 }
...
  schemas:
...
    Error:
      type: object
      required: [code, message, trace_id]
      properties:
        code:
          type: string
          enum: [SCHEMA_INVALID, UNAUTHORIZED, MISSION_NOT_FOUND, PARAM_NOT_FOUND, DOC_NOT_FOUND,
                 DOC_NOT_READY, STATE_CONFLICT, RATE_LIMITED, PROVIDER_FAILED, INTERNAL]
        message: { type: string }
        details: { type: object, additionalProperties: true }
        trace_id: { type: string }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/Error' }

    # ---------- 枚举 ----------
...
    ParameterStatus:
      type: string
      enum: [proposed, needs_review, verified, rejected]
    SourceType:
      type: string
      enum: [calc, doc, catalog, assumption, human]
    ValidationStatus:
      type: string
      enum: [pass, warn, block]
...
    GateAction:
      type: string
      enum: [accept, reject]
...
    Module:
      type: string
      enum: [system, power, ttc, data, adcs, orbit, payload, risk, other]

    # ---------- 任务 ----------
...
    ParameterSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        unit: { type: string }
        value: { type: number }
        status: { $ref: '#/components/schemas/ParameterStatus' }
        source_type: { $ref: '#/components/schemas/SourceType' }
        margin: { type: number, nullable: true }
        validation_status: { allOf: [{ $ref: '#/components/schemas/ValidationStatus' }], nullable: true }
        updated_at: { type: string, format: date-time }
    ParameterList:
      type: object
      required: [items, total, page, page_size]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ParameterSummary' } }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
    Provenance:
      type: object
      properties:
        type: { type: string, enum: [calculation, citation, assumption, human] }
        record:
          type: object
          nullable: true
          additionalProperties: true
          description: 计算记录 / 引用 / 假设 / 人工记录（结构随 type 而定，见 docs/03）
    ParameterDetail:
      allOf:
        - $ref: '#/components/schemas/ParameterSummary'
        - type: object
          properties:
            source_ref: { type: string }
            confidence: { type: string, enum: [high, medium, low] }
            version: { type: integer }
            provenance: { $ref: '#/components/schemas/Provenance' }
            validations:
              type: array
              items:
                type: object
                properties:
                  rule_id: { type: string, example: V4 }
                  status: { $ref: '#/components/schemas/ValidationStatus' }
                  created_at: { type: string, format: date-time }
    ReviewRequest:
      type: object
      required: [action, reviewer]
      properties:
        action: { $ref: '#/components/schemas/GateAction' }
        comment: { type: string, maxLength: 500, description: reject 时必填 }
        reviewer: { type: string, maxLength: 100 }
        edited_value: { type: number, description: 仅 accept 可用；修改后版本 +1 }
    ReviewResponse:
      type: object
      properties:
        param_id: { type: string }
        status: { $ref: '#/components/schemas/ParameterStatus' }
        gate:
          type: object
          properties:
            gate_id: { type: string }
            action: { $ref: '#/components/schemas/GateAction' }
            reviewer: { type: string }
            ts: { type: string, format: date-time }
        revalidation:
          type: object
          properties:
            queued: { type: boolean }
            scope: { type: string }
...